Some learning
A formula is used to generate a solve struct, but the contents of the formula are no longer used during the solve.
At some point the formula will be removed…

As variables are now stored as part of the solve, it seems the variable struct is an ok place to store extra information relating to variables.
So, introduced is a decision level which is set when a literal associated with the variable has been set.
This is used to, e.g., find the decision level of a clause.

The initial learning loop is quite basic.
Simply, resoluion is applied until there are no other literals available at the current conflict/decision level to resolve on.

Various cosmetic changes have been introduced also…