options for performing additonal analysis before backjumping
though these likely should be expanded further
e.g. to depend on the ratio of remaning clauses, etc.
as, even if there's some benefit, i doubt the quality of a conflcit can be significantly improved and as such a ratio to help avoid additionl cost might be useful
still, the functionality allows for some initial investigation, even if this doesn't seem too promising