a return to vectors
linked lists might be useful, but things were very slow
so, back to vectors for now
here, the idea is to keep the two watched literals at the start of the clause, i.e. watch a is always index 0 and watch b is always index 1
then, for updating watches all that need to be done is to start searching from the second index and swap the appropriate index

further, i wonder if something similar to benefit of pulling literals to the front can be achieved by sometimes making a few swaps
e.g. if the found index for a new a is at n, swap n with n/2 and then with 0, etc.
or, by periodically shuffling based on VSIDS, etc.