improved resolution
at the moment the code is a messy, but things seem to work ok

before, resolution relied on clauses being in order, and on creating intermediary clauses
both these things can be avoided

the idea is to create a resolution buffer, the size of all variables, and write into the buffer each resolution instance
and, when a variable is being resolved on, remove it from the buffer

this, so far, is simple

though, what's important is also figuring out when the resolved clause is asserting
to do this, the buffer beings as a copy of the current valuation, and then the observations of the current level are cleared
this allows in-palce checking of whether the literal read from the clause conflicted with the valuation, or was none
further, a simple count of when a none is assigned is sufficient to determine whether the clause is asserting, as the only way for this to be reduced is for resolution to take place, and there is exactly one resolution possibility each time
this isn't informative about which literal is asserted, but we need to go through the buffer to collect up the clause in any case, and it's easy to then spot the only literal which did not conflict with the valuation

all this is quite nice, and provides a performance boost

the very quite nice thing is to follow, as with this there's no need to keep clauses ordered
so, there's no need for a separate watch clause, and it'll be possible to easily swap remove on basic self-subsumption
and i'm also freed from what appears to be an addiction implementing basic merge/find algs for clauses, phew!

the drawback of this approach is that it requires a buffer the size of the valuation, though i think that's ok for the problems i'm interested in at the moment
and in any case, this buffer can be made part of the solve and is no more problematic than the vec for the valuation, etc.