Variables/DB and Clauses/DB revised
* Also, while sweeping, `errs` is renamed `err`

* Activity

Both variable and clause activity are now unified under `Activity`.
This simplifies and I can't think of good reason to allow distinct primitive types for the objects.

* Variables

The variables struct has been eliminated and relevant parts split into the variable database, which now holds:

- A valuation
- Previous values
- Choice indicies
- Activity
- Watch lists

… and the internal/external mappings

With this, the valuation trait has been removed, as it had no use.

The `index` method on literals has also been removed, and the `v_id` method renamed to `var` as there's no longer a distinction between variable ids/indexes and variables.
On this, `Variable` is now an alias for the underlying u32.

This reorganisation seems to have a small performance cost, though it seems worthwhile for clarity, etc.
In particular, there's far less interior mutability, which helps reasoning about things.

Stashed is an alternative take on the valuation/choice recording.

The idea was to keep a `split_vec` with the 2n/+1 encoding for even/odd variables with cells recording choice indicies.
Conceptually, this seems quite nice, as the partial valuation is implicit, though clear, and there's more information about how the valuation came to be --- naturally, given it's the combination of what are at the moment two structures.

Performance wasn't particularly good, though could maybe have been improved.
Still, the main reason for stashing and maintaining separation was that this didn't allow for the elimination of any other structure.
Specifically, the literal db would still be needed in roughly the same form it's in for easy access to implication history.
Well, so long…

The variable struct files remains.
In part for amusement, and in part as a lot of design information could be included in documentation.

* Clauses

When using the resolution buffer only literals appearing in clauses from the trail were marked as used.
Was there some reason for this?
For now, all literals are marked as used.
This leads to more iterations on a quick check, but the implementation is more efficienct.
As this is a heuristics thing, the additional iterations aren't of much concern, though if there was (or is) a reason… ?

Subsumption is now handled by the clause db, with `automatic` transfers and watch updates.

Some attempt has been made to use clauses over stored clauses where possible.

* BCP

This prevents a clean separation of the databases.
Still, I think this is okay, as things can roughly be seen in terms of maintaining cohesion amoung these databases…