cli patches and small lib changes
Unsatisfiable core functionality has been added back to the cli

Core construction works similarly to frat construction, with buffers, etc, though this time the collected information is kept in memory.
Some notes are included on how this works.

Within the library, a few changes were made:

- The frat writer takes references, so dispatches can be picked up by additional writers.
- Some additional dispatches have been added, mostly around bcp.

The cli has been reorganised a little with the various different dispatch records split into separate files within a broader module.
Some work needs to be done on comments, tho…

As an aside…
The unsat core works as expected on a few small tests, and larger tests will be added later.
Still, on a handful of SATLIB files the core isn't particularly compact.
My guess is that this is in part due to using a circular search when updating watches.
In any case, some investigation may be worthwhile…