Heap patches and a few other things
* Heap

A second draft of the heap code.
For the most part, this amounts to abstracting most of the pointer manipulation with descriptive function names.
Here, I expect the compiler to inline everything, so no cost is expected.
To pair with this, most indexing is now unchecked, as growing the heap is fine.

Somewhere, somehow, clause activity wasn't being set quite right.
This, I think, is now fixed, and there is a increase to performace in some key test cases.

The overall goal here was to decouple reduction from restarts, as the heap should track activity.
Though the fixes are sufficient for a commit…

* env_logger over log4rs

As I've found this quicker to work with…

* cli window

Now conforms to whether stats are shown, though the implementations a quick hack.
This is still a todo.

* Hash

Is no longer derived for the literal structure.
For some reason compilation was okay, but an error was thrown with clippy, which then blocked error reporting on the cli crate.
`cargo clippy --all-targets` was the final key to solving this headache…