Dispatches: second pass
The dispatch implementation is functional, though is there room for improvement?

The two primary concerns are:

A. Clear pattern matching for destructuring dispatches
B. Low overhead for structuring dispatches

Most serialisation options are ruled out by (A) or (B).
At least, with the current scope of the project.
As, … deserialisation would be required.
And, serialisation also takes work.

This includes libraries like speedy or rkyv as well as simple approaches like using a string.

It seems, then, that enums are the way to go, especially with respect to (B).
And, with (B) alone an exhaustive, flat, enum, of all the different messages would be okay, I think.
A long, but readable, dictionary.
Still, (A) pushes toward a heirarchy of enums.
There are choices here, though for now it is:

Dispatch sort => dispatch location => dispatch message

I doubt this will change, as location is optional in some cases (stats) while sort (e.g. comment, delta) is not.

The primary drawback here is expressing dispatches, as importing everything isn't so practical, there's a lot of `namespacing`.

The hope is this can be mitigated with the help of some macros.
And, more genrally the macro will allow, hopefully, disabling dispatches either at compilation or dynamically on a flag much like logs…