Optional dispatches
Through the library transmitters are optional, and all dispatch actions are scoped under the existence of some transmitter.

Tests use transmitters when needed (FRAT) and skip when not (SATLIB, etc.)

The CLI also only creates transmitters/receivers/channels when needed.
On an uninformative single-run-test the overhead of dispatches is roughly 10%.
This is roughly in line with unoptimistic expectation, at least for the particular test.
Some testing is needed to figure whether smaller or fewer messages would be an improvement…