Initial reductions between restarts
An activity heap for clauses has already been implemented, and this can be used to track which clauses are active.
Though, `activity` heap is a slight misnomer, as clauses on the heap are * inactive *.

Here, `clauses` refers only to `learnt clauses`.

To start, clauses are automatically added to the activity heap.
Then, whenever a clause is used during bcp or to force a literal (after backjumping from analysis) the clause is removed from the heap.
And, whenever a restart happens all clauses are added back to the heap.

So, between restarts no clause that has been used for bcp/forcing will be removed from the solve.

This mirrors, mostly, how literals are chosen.

On the todo list now is to figure out a good reduction strategy.
At the moment, it's a simple interval counter.
Though, it could be an increasing bound on the number of clauses learnt, etc…