It's time for a commit…
* Dependency changes

- crossterm has been replaced with termion
  extensive support for the cli isn't so important, and this reduces dependecies when compiling

- rand has been slimmed
  and, to help with this a minimal pcg implementation has been introduced
  further reductions here are unlikely, as while the dependency on rand could in principle be avoided, this wouldn't be very helpful for exploring alternatives

* Configuration

Configuration has been split up and reworked a little.
For the cli, default consts are no longer available, the motivation being cleaner documentation on the lib without.
For the lib, a few things are renamed and shuffled, and notably `sub-` configs have been added for databases and booleans.
Though, some booleans are likely to be switched to enums.

* The solve

The conflict ceremony has been reworked into a handful of more descriptive methods.

* Docs

Very rough sketches have been added for a few things.

Documenting is a somewhat unnerving.
On the one hand, there's really not *that* much here, and so much more that could be added.
On the other hand, there's quite a lot, and many things that can be improved…