Clarifications
* FRAT

Most of the code for generating FRAT proofs has been refactored.
It is, now, much more readable and some redundant actions have been removed.
Initial documentation has been added, though for the most parts it's notes to check the structure makes sense.

* Clauses and a clause trait

Clauses are now a vector of literals, and types have been updated to where a vector of literals is intended to be interpreted as a clause, but not where not.

* Config

The config has been revised a little futher, and some documentation sketches have been added.

* Functionality

Is the same as before…