Two examples, and a (temp?) patch
* Patches

The DIMACS parser made a check for a misplaced preamble.
This conflicts with using 'p' as a variable.
Though, using 'p' also conflicts with the dimacs spec, so…

I think I will likely continue to omit the extra check, as supporting a wider class of variable expressions seems nice.

* Examples

The simple example has been reworked, and an example of creating and solving a DIMACS file in memory has been added.