literal revision
* bcp source simplified to omit 'from' literal

the from info was not used, and makes generalising the use of the source to missed implications, etc, difficult
so, it has been removed

* missed implication literal source replaced with bcp

tracking missed clauses may be interesting, but can be implemented as an additional dispatch

* forced literal source replaced with bcp

as with missed implication, this source was adding additional information to an instance of bcp, which may be of interest, but the interest complicates the fundamental classification, and so should is separated

* pure literal source removed

replaced with 'choice'
this may introduce some ambiguity with dispatches, but for the moment pure literals can implicity be tracked as choices made with an empty choice stack, which are not added to the choice stack
as, if the solver is sound, any value must have been permissible for the choice to have been made

* clause key type added for unit clauses

the key is the literal itself
as learnt clauses store an index and a revision, with the index as a u32 and the revision as a u16 and literals are a variable as a u32 and a boolean, this doesn't increase the size of clause keys
this may change, though if the size is a concern i don't think think this would be too hard to identify... maybe
in any case, there no easy solution to anticipate a change while guaranteeing size, as there's no universal X to u32+u16 map, and it's not as though the current size was chosen for a particular reason anyhow...

* backjump level generalised

now takes something which implements a clause to allow for easy unit clause checking

* resolution generalised

to work with unit literals
in particular, the prior clause requirement for a literal obtained by resolution has been dropped, as the literal itself is the clause

for the moment, things have been reorganised to match the source of the literal, with most of the prior functionality now under the bcp branch, and the relevant parts from the bcp branch copied over to the resolution branch, with the literal itself substituted for the clause key

this should be revised at some point!
esp. as there's a lot of the subsumption logic in the function, which really doesn't need to be there