Scattered 'gen' and made some splits
The unified error file remains, but for more general 'Ok' results what was in the 'gen.rs' file has been moved to be close to where the corresponding thing is defined.
With this, the namespacing is then used to give a clear indication about what the enum is for, and so long as module names are descriptive, the corresponding imports are too.
E.g. the previous analysis enum is now an 'Ok' enum in 'analysis.rs' and is referenced in 'solve.rs' via analysis::Ok, etc.

The VSIDS enum, choice procedure, and applying consequences procedure, have also been split into separate files.

Some notes added in places.