generic contexts, even more documentation
- perhaps too much bundled this time…

the Context struct has been replaced with GenericContext<MinimalPCG32> and almost all impls of Context are now of GenericContext<R: Rng + Default>.
This for sure seems the way to provide a default source of randomness, though does seem to come with some small overhead.
Still, I think the overhead is ok given the flexibility secured, and this should also work as a template for making other things generic.

- context build method documentation complete
- most dispatch documentation is ok, and the rest will be done slowly, as it's quite the headache…