solve documentation, macro file
first pass at high level solve documentation, which in turn is going to require revisions to apply_consequences.
still, it seems mostly ok

to help keep things tidy a file for macros has been added within the dispatch module.
and, some macros have been simplified to not take dispatch if self is provided.

very minor changes to solve
the only maybe notable difference is a change of Ok::UnitClause to return a literal rather than a clause key.
this is to avoid ignoring the other clause key variants.