Enum disambiguation
Directly, to avoid collisions when generating C/++ header files.
Though, also in part as the overloading was a coin flip, and in hindsight…

Each enum in the err mod has been renamed following:

<Enum Name> => <Enum Name>ErrorKind

The same type of renaming has also been applied to other structures.
E.g., Reports and Oks.

Other structs will be revised when collisions are observed.