builder and database updates
Some simplification…

- Building/Parsing

A context no longer does (much) parsing.

Instead, the context provides a method to get a fresh atom, and it's up to the user of a context to keep track of what the atom represents.
This, then, removes the need for any internal/external maps, though does leave it up to the user.

The exception here is parsing DIMACS input.
Here, the existing (simple) parser is kept, but now returns a map from atoms present in the DIMACS file (as u32s) to the corresponding atom in the solver (as an Atom).

The primary motivation for this change is to keep the context (relatively) focused and simple.

- The true

Without an external mapping, there's no choice but to represent atoms as they are.
And, with this the issue of using zero (0) as an atom returns.
Well, the issue of using zero as an atom, without zero being signed.

The solution, I think, is to treat zero as true.
It's always true, and in this respect, can be ignored.
Aside from the space used, this isn't a significant change.
It amounts to adding an initial fresh clause at the creation of the atom database.

And, true rather than false as the valuation is a conjunction, rather than a clause.

- Additional methods

Without an internal/external distinction, it makes much more sense to directly interact with a context.
That is, to use native strucutres such as atoms and literals.
So, some methods have been added to help with this.

- Examples

Have been updated.
And, I think some have been improved a little.
Though, sudoku perhaps could do with a second pass.

The construction of the nonogram clauses also seems slower.
Though, that's for some other time…

- Errors

A general error kind has been introduced, and some build errors collapsed into this.

- FRAT proofs

As there's no longer a way to track the external representation of an atom, relevant parts of the transcriber have been removed.