Further IPASIR2 details
Though, for the moment this is where things will be left.

The idea was to read the API and adapt the solver in the process.
Though, some cursory reading would have been wise.

The issue, for the moment, is that unlike IPASIR1, there are many callbacks, and given these it might make some sense to update the structure of the solver.

So, as it seems the spec might be finalised soon, this will be postponed.