Revised clause implementations
Rather than a blanket implementation, there are now specific implementations for vClause and dbClause.

The motivation is to avoid unnecessary clones.
As, with the blanket deref impl, there was no way to directly return a clause already in canonical form.
So, the choice was to complicate methods which take an impl, or refine the impl.

The dbClause impl does nothing more than divert, and a clause method has been added to the dbClause to help with this.
Though, in this case using the canonical method does still result in a clone.

The construction of a nonogram is now much quicker with this.