Core updates
Back to a previous idea…

To keep track of the core, it's sufficient to store the transitive closure of bcp/resolution.
As, original clauses are never removed.
This, in turn, doesn't take too much overhead.
As, whenever a clause is derived via bcp/resolution, all that's required is the union of all the origin clauses present in the instance of propagation or resolution.

So, a first pass of this has been (re-)implemented.

For this, unit clauses now have a corresponding db clause structure, as from the previous commit.
Lookup is slow, for sure.
As, unit db_clauses are pushed to a vec, and so access requires a find.
Still, this seems preferable to complicating unit keys or allocating a vec of atom size for direct lookup.
In particular, as at the moment there's no interest in access other than for an unsat core.

The other change is passing around origins.
This uses a vec, rather than option vec, as an empty vec is sufficient for the origin clause case.

Some other minor changes, but nothing too significant.

The overhead does seem quite noticable.
So, maybe some more thinking…