Unit split
A distinction is made between original and addition unit clauses.

For the core, there's really no need to keep track of original unit clauses.
Under no circumstances (subsumption included) could these clauses have non-empty origins.
However, for FRAT proofs, it's still important to know whether an original unit clause has been added.

So…

When attempting to retreive an original unit clause from the clause db, this fails.
Unit clauses are never removed, and combined with this information the inference that the clause must be original is clear, and avoids a costly find.
Addition clauses, though, do require a find.

Otherwise, nothing else changes too much.

The core dispatch file has been removed, and the cli edited to compile, and to be refined when the core is finished.

As, for now, a puzzle is why performance is severly impacted on a few tests.