Slightly more efficient, though…
There's no need to take the transitive closure of the implication graph in all cases.
In particular, with unit clauses, these will be kept, so through BCP the only information needed to be stored is the clause used.
From this, everything else can be inferred.

For the moment, retreiving a core has not been updated with this.

It's only a minor improvement, though.
There's still something…