Reference counting
Taking the transitive closure was too expensive.

So, the only other option is to retain sufficient information to recover the implication graph.
For this, reference counting is used.
As before, the premises used to derive a clause are stored.
This time, though, only the immediate premises are kept, and a counter in each premise is updated.
Any non-zero value for the counter prevents deletion.
So, while the clause is present in the database, none of the premises used to derive the clause can be deleted.
Though, as the clause itself starts with a zero value for the counter, it may be deleted.
And, when deleted the counter of each premise used is decremented by one.

In other words, simple reference counting.

In terms of code, this is fairly minor.
Though, there's (potentially, at least) a signficant impact on heuristics, as far more clauses are locked in the database.
And, in some of the tests this shows.
Still, for most it's not too much of an impact.
And, as the restrictions imposed can easy be relaxed --- and further, the code disabled --- this seems ok.