Core update
The core procedure has been updated.
Some tests are needed, as it's a little less straightforward that with the closure.

In particular, inferring when a unit addition was from bcp or resolution is somewhat delicate, and will break if subsumption is used.

- Fixed issue with clause removal

A check on whether a leaf clause was inactive before adding to the heap wasn't made.
This had a significant impact on the qausigroup tests.
These have been split up a little, and the main impact was with 3 -- in particular -- and 5.
With the fix, things are still a little slower, but it's not too bad.

- Misc things

'Origins' have been renamed to 'premises'.
The 'Source' kind in the clause module has been renamed to 'ClauseSource'