IPASIR Updates
Resets are handled as well as can be, I think.

Roughly, there's no API call for a reset, so they must be handled implicitly.
And, it seems easiest to do so with checks at relevant function calls.

For failed literals, some additional structures are added.
And, these are cleared on a new solve.

Termination callback has been added, though in non-final form.
With this, though, things should be suffient to build a few tests.