IPASIR callback structure, and the learn callback
The structure bundles all the ipasir relevant callback requirements, and whether a callback should be made can be determined with two checks.
This seems reasonable, and avoids the presence of dangling pointers unless needed.

The learn callback is still todo, as it does force a transparent representation of atoms, somewhat…