Support for ensuring transparent representation of atoms
This is fairly straightforward.
As, all that's required is to add as many atoms as needed until a given natural is hit.

The direct use of this is removing the internal/external mapping used for the IPASIR API.

The concern is bound variable elimination and other procedures which may lead to some misalignment.
Though, it seems preferable to consider implementation issues at the time of implementation.
That is, in part, the purpose of this…

To illustrate the ensuring atoms, the assumption example has been updated.