A few more tests, and a few fixes
All the remaining IPASIR examples which compile with reasonable effort have been added.

These relevealed some issues:

- The unsat core cache was not being cleared when needed
- The method of detecting an unsat formula while building doesn't suit the IPASIR API

Roughly, the issue is the API doesn't require an explicit call to refresh the solve.

So, some inference needs to happen.
The problem is, this inference should run when an assumption happens, but then without additional state it's unclear whether an unsat state is due to a previous solve to be cleared, or due to some prior assumption already entailing unsat, etc.

The fix for this, for the moment, at least, is to assert all assumptions and then check all clauses at the start of a solve.

Aside from these two things, things are mostly ok.

Though, two required fixes are introduced:

- Undoing assumptions is unsound, if an assumption coincides with a unit clause.
- On an immediate conflict the conflict key is a placeholder.