Default to int based literals & clauses w/ cfg for boolean
Given the IPASIR API, a way to avoid casts is useful.
And, without much surprise, using ints is quicker.
Though, this may be in part due to size, as ab literals are larger, as u32 + byte vs i32.
Strictly, ab lits could be u16, but this requires a lot of casts, and so hasn't been examined.

The cfgs are a little messy, but using the cfg! macro resulted in compile errors due to conflicting closure types in some cases…