Some IPASIR documentation
Importantly, a todo as a unknown solve should return to input.

Otherwise, some documentation of things in need of documentation.
Mostly, IPASIR API.

Otherwise, i_clause has been updated to int_clause.

Of note:

An inital effort to make the IPASIR module and related context parts conditional on a (default) feature were made, but will not bw committed.
Simply, as the tooling I use would then make working with the IPASIR source much more difficult, and it would be more difficult to maintain.
And, in any case, performance is hardly affected.