Callback refinements and some IPASIR2 details
An intermediate commit.

The significant change is that callbacks are now unified under a single structure with a version field.
And, callbacks are now made through methods to the unified structure.

The change has been made to help reason about the callbacks, as the v2 API will introduce some complexity beyond storing additional fixed fields.
Specifically, the learn/import callback takes an additional argument in v2.

As a result, the callback struct is now passed through storing a clause.

Otherwise, a fix:

- Learn callbacks are made for unit clauses

And some additions:

- IPASIR2 functions which extend IPASIR functions use the IPASIR function where possible.
- Adding a clause via IPASIR2 now uses the clause buffer.