IPASIR2 Callback approximation
Fields have been added for each IPASIR2 callback, and all but import have at least partial implementations.

For import, a note has been made on the need for additional state and will be explored at some later time.
In this respect, add is also partial, as functionality for addition during a solve has not been implemented.
Though, neither has the option to mark a clause as optional.

For fixed, further refinements to assumptions need to be made before this can be finalised.
Though once a clear abstraction for the bottom level is implemented it should be straightforward.