Foundations for 'stacked' assumptions
Prior to this commit, all assumptions were bundled on a distinct decision level.
This commit adds initial support for 'stacking' assumptions as though they were decisions.

This is motivated by an option to choose between the two in the IPASIR2 draft.

Conceptually, the change is simple:

Assumptions are added exactly as decisions, and a variable is used to keep track of where levels transition from assumptions to decisions, to ensure backjumping does not clear an assumption.

In practice, this requires untangling two things (the lowest decision level, and the inital level) which was somewhat delicate.
Though, a few literal_db methods have been improved due to this.

What's missing from this commit is processing the assumptions in sequence.
That is, at the moment a level is added for each assumption, but all consequences are pushed to the 'top' level.

A literal_db config has been added.
Though, it might be worthwhile revisiting the way configs are handled…