Flat assumption bugfix, lsp test update
I had assumed there would be no problem setting top as the assumption when handling all assumptions together.
This, with a little reflection, is not the case.

So, in place of this, some arbitrary assumption is chosen as the designated assumption.
After this, every other assumption is queued, rather than being handled by a specific instance of bcp.
As, this should be equivalent and is more in keeping with intent.

Otherwise, for stacked assumptions, the decision level is now anticipated, and so should be correct.
Not that this has any effect in practice…

The longest simple path test has been revised in line with other tests.
Checking this test after switching back to flat assumptions highlighted the problem.