Assumption fix, reachability test revised
With the revisions for stacked assumptions, clearing non-stacked assumptions was overlooked.
So, this commit restores dropping the value of every recorded assumption when clearing assumptions.

With this, the previous fix is likely redundant.
As, with the lsp tests a single assumption is made, and adding this as the representative assumption for a level ensured the assumption was cleared appropriately.
Still, I think keeping some representative is preferable to using top, as it means the non-stacked assumption does not need to be handled as a special case.