Updated readme example, and a helpful method
The all_models example has been refreshed to be more in line with the library as it stands.
And, the same example has been updated in the readme.

In addition, the DIMACS example has been updated, and the redundant example removed.

Added is a method to receive an iterator over all atoms and values, where the atom has some value.
This helps skip a fair amount of matching.