Resolution simplified and transfer removed
Transfer of a long clause to binary was quite neat.
However, as this involved a change in clause keys it is incompatible with the current approach to obtaining an unsat core.

So, some options:

1. Revise the core (again)
2. Make certain features conditional
3. Remove subsumptions.
4. Remove transfer and allow some binary clauses to be stored as long.

The last option seems preferable.
Subsumption is now a quite simple feature.
And, at a high level should remain simple.
Roughly, in some cases accumulated premises from resolution are used to strengthen a clause.
That's it, and more complex instances of subsumption (someday…) should also fit this description.

Additional upsides are maintainability and allocs for binary clauses.
The only downside which comes to mind is that the heuristic for processing binary clauses first will not be applied to subsumed clauses.
This seems okay.

The resolution loop has been simplified with this.
And, in addition, the FirstUIP option now only breaks the loop early.
With the above, this means there's a unique exit point from a merge call.
Further, all dispatches are now grouped.

Some complications:

With subsumption a the premise relation may have loops.
Though, rather than introduce loops explicitly, only the proof count is incremented.

And, there may be an issue with the valuation used to re-initialise watches.
That will be handled as a general update to watch handling.