Persistent resolution buffers
The resolution buffer is stored as part of the context.

The primary motivation is to allow for an easier time with callbacks.
The remaing motivation is to finally get around to this.
The difficulty is updating the buffer to suit a valuation.
This is done with the very simple approach of copying a valuation if it's smaller, and a loop if equal.
A valuation shouldn't shrink, so a todo is left for the remaining case.

For convenience, a method to identify the size of a valuation has been added.

There are no other significant changes.
Though, a holder for a callback and some related init methods have been added for the buffer config.