Resolution buffer callbacks, dispatches removed
The transient module has been renamed to resolution_buffer, given the updates made in the last commit.

Otherwise, fairly straightforward.

A callback has been added to the buffer, and as with other callbacks, the code from the macro has been moved to a closure.

Though, in this case, removing the dispatch was also possible and so has been done.