Removed the crossbeam requirement for tests
This also goes some way to simplifying most of the dispatch syntax.
For, FRAT proofs are now written directly with a transcriber, ready for the dispatch parameters to be generalised.

The CLI is reduced further, again, no particular interest here at the moment.

And, that's mostly all there is.
Performance is more-or-less the same.