CLI temporarily iced, dispaches tidied further
With the change to dispatches, the CLI does very little.
And, rahter than refactor in parallel, it seems preferable to put it on ice for the moment and return for a rewrite after some revisions have taken place.
In particular, it would be nice to handle most of the argument parsing via a revised configuration that supports IPASIR options.
In any case, that's for some other time…

Dispatches have been removed more generally.
The dispatch module has been renamed 'reports' and the SolveReport enum is now the Report enum.

There are no non-refactor related changes.