Callbacks re-organised
Callbacks can be (and are now) set for a context.
These are all grouped in a callback submodule of the context, along with other callback related things.

Callback templates have been added to help clarify the generation of FRAT proofs.
An example is to be added soon…

Some redundant documentation has been removed.

The TimeUp variant of report has been removed, to help keep things simple.