Configuration refinements
A ConfigOption struct has been defined and the top level configuration has been updated to use this struct.

Taking insparation from the IPASIR2 API (and with an eye to easy interpolation) the struct supports setting a name, min and max values, the maximum state of the context when setting, and the value.
Other fields should be fairly simple to add when needed.

The struct is generic over the underlying config type.
So, to help maintain things custom underlying structs (VSIDS, etc.) have been given set min/max values.

For the moment, all fields are accessed directly, though this may change.

The only significant change is removing the optional from the time limit.
Here, zero is sufficient to indicate no time limit, as if so there would be no need to call the solve.