Further config refinements
ConfigOption is now used for all the config options.

The Scheduler struct has been removed, and scheduling options moved to the top level.
Further, these are no long optional, as the max value is sufficent, I think.
The upshot is that checks are much more simple.

In addition, the duplication stacked_assumptions config has been removed.